KIEV21DEVS-123 Add mapper, checkstyle plugin and refactoring.
