+++ src/main/java/com/heroku/spacey/controllers/AuthController.java
@@ -5,11 +5,10 @@ import com.heroku.spacey.dto.user.UserRegisterDto;
 import com.heroku.spacey.dto.auth.UserTokenDto;
 import com.heroku.spacey.models.UserModel;
 import com.heroku.spacey.security.JwtTokenProvider;
-import com.heroku.spacey.services.MailServiceImpl;
-import com.heroku.spacey.services.TokenService;
-import com.heroku.spacey.services.UserService;
+import com.heroku.spacey.services.impl.MailServiceImpl;
+import com.heroku.spacey.services.impl.TokenService;
+import com.heroku.spacey.services.impl.UserService;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
