+++ src/main/java/com/heroku/spacey/mapper/CategoryForProductMapper.java
@@ -1,4 +1,4 @@
-package com.heroku.spacey.dao.category;
+package com.heroku.spacey.mapper;
 
 import com.heroku.spacey.entity.Category;
 import org.springframework.jdbc.core.RowMapper;
@@ -6,12 +6,13 @@ import org.springframework.jdbc.core.RowMapper;
 import java.sql.ResultSet;
 import java.sql.SQLException;
 
-public class CategoryMapper implements RowMapper<Category> {
+public class CategoryForProductMapper implements RowMapper<Category> {
+
     @Override
     public Category mapRow(ResultSet resultSet, int i) throws SQLException {
         var category = new Category();
-        category.setId(resultSet.getInt("id"));
-        category.setName(resultSet.getString("name"));
+        category.setId(resultSet.getInt("category_id"));
+        category.setName(resultSet.getString("category_name"));
         return category;
     }
 }
