+++ src/main/java/com/heroku/spacey/mapper/CategoryMapper.java
@@ -1,4 +1,4 @@
-package com.heroku.spacey.dao.product;
+package com.heroku.spacey.mapper;
 
 import com.heroku.spacey.entity.Category;
 import org.springframework.jdbc.core.RowMapper;
@@ -6,13 +6,12 @@ import org.springframework.jdbc.core.RowMapper;
 import java.sql.ResultSet;
 import java.sql.SQLException;
 
-public class CategoryForProductMapper implements RowMapper<Category> {
-
+public class CategoryMapper implements RowMapper<Category> {
     @Override
     public Category mapRow(ResultSet resultSet, int i) throws SQLException {
         var category = new Category();
-        category.setId(resultSet.getInt("category_id"));
-        category.setName(resultSet.getString("category_name"));
+        category.setId(resultSet.getInt("id"));
+        category.setName(resultSet.getString("name"));
         return category;
     }
 }
