+++ src/main/java/com/heroku/spacey/security/JwtAuthenticationFilter.java
@@ -1,6 +1,6 @@
 package com.heroku.spacey.security;
 
-import com.heroku.spacey.services.UserService;
+import com.heroku.spacey.services.impl.UserService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpHeaders;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
