+++ src/main/java/com/heroku/spacey/services/MailService.java
@@ -1,4 +1,4 @@
-package com.heroku.spacey.contracts;
+package com.heroku.spacey.services;
 
 public interface MailService {
     void sendSimpleMessage(String to, String subject, String text);
