+++ src/main/java/com/heroku/spacey/services/impl/MailServiceImpl.java
@@ -1,6 +1,6 @@
-package com.heroku.spacey.services;
+package com.heroku.spacey.services.impl;
 
-import com.heroku.spacey.contracts.MailService;
+import com.heroku.spacey.services.MailService;
 import lombok.SneakyThrows;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
