+++ src/main/java/com/heroku/spacey/utils/CategoryConvertor.java
@@ -1,4 +1,4 @@
-package com.heroku.spacey.mapper;
+package com.heroku.spacey.utils;
 
 import com.heroku.spacey.dto.category.CategoryDto;
 import com.heroku.spacey.entity.Category;
