+++ src/main/resources/sql/category_queries.properties
@@ -10,4 +10,4 @@ update_category=      UPDATE categories \
                       WHERE id = ?
 
 delete_category=      DELETE FROM categories \
-                      WHERE id=?
+                      WHERE id = ?
