+++ src/main/resources/sql/material_queries.properties
@@ -10,4 +10,4 @@ update_material=       UPDATE materials \
                        WHERE id = ?
 
 delete_material=       DELETE FROM materials \
-                       WHERE id=?
+                       WHERE id = ?
