+++ src/main/resources/sql/product_detail_queries.properties
@@ -10,4 +10,4 @@ update_product_detail=      UPDATE product_details \
                             WHERE id = ?
 
 delete_product_detail=      DELETE FROM product_details \
-                            WHERE id=?
+                            WHERE id = ?
