+++ src/main/resources/sql/product_queries.properties
@@ -27,7 +27,7 @@ update_product=               UPDATE products \
                               WHERE id = ?
 
 delete_product=               DELETE FROM products \
-                              WHERE id=?
+                              WHERE id = ?
 
 deactivate_product=           UPDATE products \
                               SET isavailable = false \
